@@ -4,16 +4,16 @@
   "author": "Raymond-Yip",
   "private": true,
   "engines": {
-    "node": ">= 0.12.3",
-    "npm": ">= 2.9.1"
+    "node": ">=4.1",
+    "npm": ">=3.1 <4"
   },
   "dependencies": {
-    "eventemitter3": "1.1.0",
     "fastclick": "1.0.6",
-    "flux": "2.0.3",
     "lodash": "3.9.3",
     "normalize.css": "3.0.3",
-    "react": "0.13.3"
+    "react": "^0.14.3",
+    "react-addons-css-transition-group": "^0.14.3",
+    "react-dom": "^0.14.3"
   },
   "devDependencies": {
     "autoprefixer-core": "^5.2.0",
@@ -21,14 +21,12 @@
     "babel-loader": "^5.3.2",
     "bluebird": "^3.0.5",
     "browser-sync": "^2.7.7",
-    "coffee-jsx-loader": "^0.1.2",
-    "coffee-loader": "^0.7.2",
-    "coffee-script": "^1.9.3",
     "css-loader": "^0.14.5",
     "csscomb": "^3.1.7",
     "cssnext": "^1.8.3",
     "del": "^1.2.0",
     "extract-text-webpack-plugin": "^0.8.2",
+    "file-loader": "^0.8.5",
     "gaze": "^0.5.1",
     "jest-cli": "^0.4.18",
     "less": "^2.5.1",
