@@ -1,14 +1,66 @@
-import React, { PropTypes } from 'react';
+import React, { PropTypes, Component } from 'react';
 import PageComponent from '../page';
 import NavComponent from '../nav';
-import styles from './app.less';
 import DropdownComponent from '../dropdown';
+import './app.less';
 
-const App = React.createClass({
-  displayName: 'App',
-  mixins: [],
-  render: function(){
+export default class App extends Component {
+  state = {
+    pageIndex: 0,
+    status: 0
+  }
+
+  setStyle = (val, index) => {
+    for (var prop in val) {
+      if (val.hasOwnProperty(prop)) {
+        this.state.pageData.pages[this.state.pageIndex].cmps[index].style[prop] = val[prop];
+      }
+    }
+    this.setState({
+      pageData: this.state.pageData
+    });
+  }
+
+  prevPage = (e) => {
+    let nextIndex = this.state.pageIndex;
+    if (nextIndex > 0){
+      nextIndex--;
+    }
+    this.setState({
+      pageIndex: nextIndex
+    });
+  }
+
+  nextPage = (e) => {
+    let nextIndex = this.state.pageIndex;
+    if (nextIndex < this.state.maxPageIndex){
+      nextIndex++;
+    }
+    this.setState({
+      pageIndex: nextIndex
+    });
+  }
+
+  loadPageData(){
+    let data = require('../../lib/js/data');
+    this.setState({
+      maxPageIndex: data.data.pages.length - 1,
+      status: 1,
+      name: data.data.name,
+      pageData: data.data
+    });
+  }
+
+  componentDidMount(){
+    if(this.state.status == 0){
+      this.loadPageData();
+    }
+  }
+
+  render(){
+
     let PAGECOMPONENT = '';
+
     if(this.state.status){
       let item = this.state.pageData.pages[this.state.pageIndex];
       $('#app').css('background', item.bgcol);
@@ -19,60 +71,13 @@ const App = React.createClass({
       setStyle = {this.setStyle}
       showTitle = {false}/>
     }
+
     return (<div className='App-box' ref='APP' {...this.props}>
       {PAGECOMPONENT}
       <NavComponent nextPage={this.nextPage} prevPage={this.prevPage}/>
     </div>
     );
-  },
-  setStyle: function(val, index){
-    for (var prop in val) {
-      if (val.hasOwnProperty(prop)) {
-        this.state.pageData.pages[this.state.pageIndex].cmps[index].style[prop] = val[prop];
-      }
-    }
-    this.setState({
-      pageData: this.state.pageData
-    });
-  },
-  prevPage: function(){
-    let nextIndex = this.state.pageIndex;
-    if (nextIndex > 0){
-      nextIndex--;
-    }
-    this.setState({
-      pageIndex: nextIndex
-    });
-  },
-  nextPage: function(){
-    let nextIndex = this.state.pageIndex;
-    if (nextIndex < this.state.maxPageIndex){
-      nextIndex++;
-    }
-    this.setState({
-      pageIndex: nextIndex
-    });
-  },
-  loadPageData: function(){
-    let data = require('../../lib/js/data');
-    this.setState({
-      maxPageIndex: data.data.pages.length - 1,
-      status: 1,
-      name: data.data.name,
-      pageData: data.data
-    });
-  },
-  componentDidMount: function(){
-    if(this.state.status == 0){
-      this.loadPageData();
-    }
-  },
-  getInitialState: function(){
-    return {
-      pageIndex: 0,
-      status: 0
-    };
   }
-});
 
-export default App;
+}
+
