@@ -1,30 +1,29 @@
-import React, { PropTypes } from 'react';
-import Touch from '../../utils/touch';
-import styles from './image.less';
+import React, { PropTypes, Component } from 'react';
+import './image.less';
 
-const Image = React.createClass({
-  displayName: 'Image',
-  propTypes: {
+export default class Image extends Component{
+  static propTypes = {
     url: PropTypes.string.isRequired,
     style: PropTypes.object
-  },
-  mixins: [Touch],
-  render: function(){
+  };
+
+  static defaultProps = {
+    type: 'image'
+  };
+
+  state = {
+    cls: ['image', 'j-component']
+  };
+
+  render() {
     return (
       <img
       src={this.props.url}
       className={this.state.cls.join(' ')}
       style={this.props.style}/>
     );
-  },
-  getDefaultProps: function(){
-    return {type: 'image'};
-  },
-  getInitialState: function(){
-    return {
-      cls: ['image', 'j-component']
-    }
   }
-});
 
-export default Image;
+};
+
+// <ImageComponent url style/>
