@@ -1,16 +1,39 @@
-import React, { PropTypes} from 'react';
+import React, { PropTypes, Component} from 'react';
 import ModalComponent from '../modal';
 import FontModalComponent from '../fontmodal';
-import styles from './nav.less';
 import utils from '../../utils/utils';
+import './nav.less';
 
-const Nav = React.createClass({
-  displayName: 'Nav',
-  render: function(){
+export default class Nav extends Component {
+  state = {
+    isModalOpen: false,
+    modalMain: ''
+  };
+
+  openModal = () => {
+    this.setState({
+      isModalOpen: true
+    });
+  }
+
+  closeModal = () => {
+    this.setState({
+      isModalOpen: false
+    });
+  }
+
+  createFont = () => {
+    this.state.modalMain = (
+      <FontModalComponent clickCb={this.closeModal}>
+      <button onClick={this.closeModal} className="btn btn-danger">取消</button>
+      </FontModalComponent>);
+    this.openModal();
+  }
+
+  render() {
     return <div>
       <ul className="nav">
         <li onTouchStart={this.createFont}>Text</li>
-        <li>Image</li>
         <li onTouchStart={this.props.prevPage}>Prev</li>
         <li onTouchStart={this.props.nextPage}>Next</li>
       </ul>
@@ -20,30 +43,6 @@ const Nav = React.createClass({
       {this.state.modalMain}
       </ModalComponent>
     </div>
-  },
-  getInitialState: function(){
-    return {
-      isModalOpen: false,
-      modalMain: ''
-    };
-  },
-  openModal: function(){
-    this.setState({
-      isModalOpen: true
-    });
-  },
-  closeModal: function(){
-    this.setState({
-      isModalOpen: false
-    });
-  },
-  createFont: function(){
-    this.state.modalMain = (
-      <FontModalComponent clickCb={this.closeModal}>
-      <button onClick={this.closeModal} className="btn btn-danger">取消</button>
-      </FontModalComponent>);
-    this.openModal();
   }
-});
 
-export default Nav;
+};
